### 2016-05-30 #25 @whedon
Submission: GeneNetwork: framework for web-based genetics
**Submitting author:** @pjotrp (<a href="http://orcid.org/0000-0002-8021-9162">Pjotr Prins</a>)
**Repository:** <a href="https://github.com/genenetwork/genenetwork2" target ="_blank">https://github.com/genenetwork/genenetwork2</a>
**Version:** 2.0-a8fcff4
**Archive:** <a href="http://dx.doi.org/10.5281/zenodo.53740" target ="_blank">10.5281/zenodo.53740</a>

- [x] **Version:** Does the release version given match the GitHub release (2.0-a8fcff4)?
- [x] **Archive:** Does the <a target="_blank" href="http://dx.doi.org/10.5281/zenodo.53740">software archive</a> resolve?

- [ ] **Performance:** Have the performance claims of the software been confirmed?

- [x] **Functionality documentation:** Is the functionality of the software documented to a satisfactory level (e.g. API method documentation)?
