### 2016-05-31 c39e0fb
Allow the json dir to be in the old location.
