### 2016-05-31 6379feb
Renamed TEXTDIR to GENERATED_TEXT_DIR
