### 2016-05-31 d937689
Fixed error when MrnaAssay search results didn't have a geneid
