### 2016-06-01 #25 @krother
This is how software papers I like to read should look like!
I have a number of small suggestions:
- 1st par: "experimental data" could be fleshed out in 2-3 words for the biologist reader (is it SNPs, microarray data, NGS, Y2H or manually counted individuals?)
- 2nd par: "can be found in ___ guix-bioinformatics", please insert 'the article', 'the package', 'the lab fridge labeled'.. whichever is aproppriate ;-)
- last par: "Python" is spelled with a capital 'P'
- REST API looks like a very good idea.
