### 2016-06-01 #25 @pjotrp
Thanks @krother for reviewing this very complex project! We'll clarify the text.

Also @agarie promised to review. Do we need a second reviewer?
