### 2016-06-01 #25 @krother
Not necessarily. I think a few of the points above (contribution, mini-summary) could be spelled out in the README.

Apart from that, I'm looking forward to see the software in action. At the moment, I'm trying to install guix (binary tarball is 404, Github issue pending).
