### 2016-06-01 #25 @pjotrp
We can support you on the #genenetwork IRC channel if you like. Guix binary reproducible installs can be a bit daunting the first time because there are quite a few newish concepts.
