### 2016-06-01 #25 @agarie
Ooops, sorry for not posting before, I'm reviewing this.

* You could add a CONTRIBUTING.md file, as explained in [Setting guidelines for repository contributors](https://help.github.com/articles/setting-guidelines-for-repository-contributors/). Specifically, what are your requirements re: testing and documentation?
* How do I run the tests?
* As @krother said, these points could be added to the README.
* I think you could move that huge IRC log to another file inside `doc/` and leave some highlights or brief quotes in the installation docs.

I read the installation instructions on [doc/README.org](https://github.com/genenetwork/genenetwork2/blob/master/doc/README.org), but I'm on a MacBook right now. I'll try to install GN2 on my linux desktop at home tonight and report back my impressions and suggestions about the process.
