### 2016-06-01 f0aadf8
Changed global gene search to no longer use server-side processing
