### 2016-06-02 e241d27
Paper: addressing https://github.com/openjournals/joss-reviews/issues/25
1st par: "experimental data" could be fleshed out in 2-3 words for the biologist reader (is it SNPs, microarray data, NGS, Y2H or manually counted individuals?)
