### 2016-06-02 9669b09
Paper: addressing https://github.com/openjournals/joss-reviews/issues/25
2nd par: "can be found in ___ guix-bioinformatics", please insert 'the article', 'the package', 'the lab fridge labeled'.. whichever is aproppriate ;-)
