### 2016-06-02 3788a3a
Paper: addressing https://github.com/openjournals/joss-reviews/issues/25
last par: "Python" is spelled with a capital 'P'
