### 2016-06-02 #25 @pjotrp
@krother I have addressed these in the following commits:

> 1st par: "experimental data" could be fleshed out in 2-3 words for the biologist reader (is it SNPs, microarray data, NGS, Y2H or manually counted individuals?)

https://github.com/genenetwork/genenetwork2/commit/e241d27ccd3302632fec52043fc8b42125f4cfd7

> 2nd par: "can be found in ___ guix-bioinformatics", please insert 'the article', 'the package', 'the lab fridge labeled'.. whichever is aproppriate ;-)

https://github.com/genenetwork/genenetwork2/commit/9669b09bdd8885acae07e05cb878e575b1b2c6e1

> last par: "Python" is spelled with a capital 'P'

https://github.com/genenetwork/genenetwork2/commit/3788a3ae75d8260c167818786aeeeb20b22fb305

> REST API looks like a very good idea.

Yes, the code is being developed in this repo: https://github.com/genenetwork/gn_server
