### 2016-06-02 e3fc8d1
Paper: addressing JOSS requirements for contribution, where to find        documentation and testing materials
https://github.com/openjournals/joss-reviews/issues/25
