### 2016-06-02 #25 @pjotrp
@krother, @agarie I think I addressed most JOSS requirements by updating the README with:
https://github.com/genenetwork/genenetwork2/commit/e3fc8d16cff08ee9f42bf937121f7119d2670ded

That includes a statement of work, contact and how to contribute. We do not prescribe guidelines for coding standards etc. Maybe that will come in the future.

Note: the testing framework is new and still limited in scope and @Brainiarc7 is working on that. We want to achieve full coverage in the coming month(s).

Note: the user level documentation is mostly from GN1 days. We need to update them for GN2, this is a slow work in process. Meanwhile the GN1 docs are still usable.

Note: Example use of GN can be found in the many publications by users mentioning GN on pubmed and Google Scholar. We never had a publication on the software itself that did justice to the work of the contributing developers.
