### 2016-06-02 #25 @krother
I find the README.md much more informative now. Thanks!
