### 2016-06-02 a4fd21d
Fixed some issues with global search result tables, including: - Wrong digits for mean expression and additive effect - Bad loooking width on some columns, like Max LRS Location - Janky functionality for column resizing and drag-and-dropping (just disabled these functions) - Geno and Pheno global search tables weren't built in the same way, so now they are
I also fixed the sorting to always put N/A's at the bottom of the table by editing naturalSort.js
