+++ wqflask/wqflask/static/new/packages/DataTables/js/dataTables.naturalSort.js
@@ -21,10 +21,12 @@ function naturalSort (a, b) {
         // numeric, hex or date detection
         xD = parseInt(x.match(hre)) || (xN.length != 1 && x.match(dre) && Date.parse(x)),
         yD = parseInt(y.match(hre)) || xD && y.match(dre) && Date.parse(y) || null;
+
     // first try and sort Hex codes or Dates
     if (yD)
         if ( xD < yD ) return -1;
         else if ( xD > yD )  return 1;
+
     // natural sorting through split numeric strings and default strings
     for(var cLoc=0, numS=Math.max(xN.length, yN.length); cLoc < numS; cLoc++) {
         // find floats not starting with '0', string or 0 if not defined (Clint Priest)
@@ -40,16 +42,26 @@ function naturalSort (a, b) {
         if (oFxNcL < oFyNcL) return -1;
         if (oFxNcL > oFyNcL) return 1;
     }
+
     return 0;
 }
  
 jQuery.extend( jQuery.fn.dataTableExt.oSort, {
     "natural-asc": function ( a, b ) {
-        return naturalSort(a,b);
+        // first check if null or n/a
+        if (a == "N/A" || a == "NA") return 1;
+        else if (b == "N/A" || b == "NA") return -1;
+        else {
+            return naturalSort(a,b);
+        }
     },
  
     "natural-desc": function ( a, b ) {
-        return naturalSort(a,b) * -1;
+        if (a == "N/A" || a == "NA") return 1;
+        else if (b == "N/A" || b == "NA") return -1;
+        else {
+            return naturalSort(a,b) * -1;
+        }
     }
 } );
  
