### 2016-06-03 #25 @agarie
The improved README is good!

I tried to build everything but couldn't download the tarball for `AnnotationDbi`. After reading that IRC log, the best way to install GN2 is apparently to download the substitutes from http://guix.genenetwork.org:8080, right? Given the complexity of the software, I'm fine with needing a bit of help during installation, but if you could put a public key on the docs ("gpg add this key to use `substitute-urls` for a binary installation"), that'd be great.

I navigated through the live site in order to experiment with the platform and I think both functionality and performance are OK. The documentation and the other checks are also OK. I would say that the API-level docs are sparse, but the software is meant to be used on a browser, not as a library. Even then, this documentation can come in handy for new contributors and anyone experiencing problems (or just being curious), but this is mostly a suggestion for the coming months, not a necessary revision for acceptance.

The JOSS reviewer guidelines explicitly say that reviewers are expected to be able to install the software. If my understanding of guix is correct, I expect to be able to install GN2 after I get a key for the caching server (`guix archive --authorize`, right?); can you send that key to me? :)
