### 2016-06-03 ab133cc
Changed appearance of tables to more easily differentiate them from the rest of the page (still not totally pleased with them though)
Fixed issue where some N/A lrs scores and additive effects were being changed to 0's, which was misleading and kept them from always being sorted at the bottom. They now appear as N/A.
