### 2016-06-03 3bc5e36
Made a bunch of changes to correlation and search result tables (mainly the former) and made all tables a little more compact by lessenning the padding
