### 2016-06-04 #25 @agarie
I got genenetwork2 working on my own machine, using the test database (linked from [here](https://github.com/genenetwork/genenetwork2/blob/master/doc/README.org#run-mysql-server)), so the installation instructions are correct. :)

With that out of the way, @arfon, my recommendation is for acceptance.
