### 2016-06-06 #25 @krother
Yes I'd like to give it a try as well.
