### 2016-06-07 f339674
R/qtl pair scan now works again, though it takes a while (10-15 minutes).
Added pop-up warning user about run-time for pair scan

Fixed links to GN2 github on the index page

Fixed a couple issues with the search result page text

Changed the appearance of the trait sample data table to match other tables (darker headers, etc)
