### 2016-06-08 #25 @pjotrp
@krother I wrote a 'quick installation recipe' which does not require building the software.

See https://github.com/pjotrp/genenetwork2/tree/paper/doc#quick-installation-recipe
