### 2016-06-08 #25 @pjotrp
BTW we also have a plant database available now (3GB). I should add that to the paper.
