### 2016-06-09 #25 @pjotrp
To get an idea of the complexity of GN2, see the graph of dependencies at http://biobeat.org/gn2.svg
