### 2016-06-09 47cf69f
Docs: add SVG graph
