### 2016-06-09 ed48249
Doc: moved graph into README
