### 2016-06-09 86e6706
README: capitalization
