### 2016-06-10 #25 @pjotrp
@krother I updated the docs and reran the installation with success in a fresh and basic Debian VM. If you follow the instructions it takes a few minutes of keyboard actions (copy-paste only). Please try
https://github.com/genenetwork/genenetwork2/tree/master/doc#quick-installation-recipe

Downloading the software takes an hour on my 5 Mbs connection. Mind, the download activity is mesmerizing, so you may spend more time looking at the scrollling screen than is considered healthy ;).
