### 2016-06-13 #25 @pjotrp
@krother you can ignore the lcale error. It is harmless.

Your system is safe because guix itself and its packages are installed in /gnu/store - which is isolated from the rest of you system. A database is installed in /var/guix and an access key in /etc/acl.

One of the great things about guix is that it runs on top of any existing system.

Even so, you can install in a VM, such as KVM or VirtualBox. We do that all the time.

I would suggest to continue where you are now. If you come on IRC we can help you. Use https://webchat.freenode.net/ and use the #genenetwork channel. You can use any nick name. We are there.
