### 2016-06-13 #25 @roelj
If I may jump in..

@krother:
Your system is safe.  The only thing Guix does, is install things to `/gnu/store` and to `/var/guix/`.  By default, commands you type are looked for in directories specified by a `PATH` variable. Guix creates a system on top of a system, so to use programs installed with Guix, you have to add an entry to the `PATH` variable. Removing it, and removing `/gnu/store` and `/var/guix` brings you back to the original state of your system.

The locale error is not a problem actually. Some special characters may not display in the output of programs, which does not hinder the actual programs (it's only an output-to-shell problem). You could fix that by updating the `LANG` and `LC_ALL` environment variables after installing locales with Guix (Guix programs are self-contained).

I am also in the #genenetwork channel on freenode.net, available to help you there.
