### 2016-06-13 #25 @krother
Thanks for the quick reply. I'll try VirtualBox .. have been looking for an excuse to try that for some time :-)
