### 2016-06-13 #25 @pjotrp
In general a wise idea when installing and testing software from other parties, though you are safe with Guix. If you get stuck I'll try to document those VirtualBox steps too.
