### 2016-06-13 #25 @arfon
Hi @krother, given that this installation seems to be causing significant trouble for you and that @agarie has managed to install the application already I'm happy to stop here if that makes things easier?

If you agree, we can (I believe) move forward to accept.
