### 2016-06-13 #25 @arfon
> Yes, I have seen a lot improving, and in my opinion the review process has worked well. I just checked the examples in the documentation and ticked the two remaining boxes in 'Documentation'. I agree with your suggestion to move forward to accept.

👍 thanks @krother.

@pjotrp - do you want to make a new Zenodo release with your changes or has this already been done?
