### 2016-06-13 #25 @pjotrp
@krother thanks!

@arfon no changes in the release - we only added to the documentation to clarify the installation process. Happy to go public on the current Zenodo doi.
