### 2016-06-14 #25 @arfon
👍 thanks for the review @krother & @agarie.

@pjotrp - your DOI is http://dx.doi.org/10.21105/joss.00025 :zap: 🚀 :boom:
