### 2016-06-16 #25 @pjotrp
@arfon thanks for the great journal! In the paper there is a typo in Prof. Wiliams' name and I amended a URL to point to the master branch. Can you fix that, please?

See https://github.com/genenetwork/genenetwork2/commit/75c7a0828625e3aaf6326e91fadf20df96207fc3
